mqtt/requirements.txt
Django==2.2.16
djangorestframework==3.11.1
importlib-resources==3.0.0
paho-mqtt==1.5.0
psycopg2==2.8.6
PyQt5==5.15.1
PyQt5-sip==12.8.1
PyQtWebEngine==5.15.1
pytz==2020.1
pywebview==3.2
sqlparse==0.3.1
zipp==1.2.0
